Move 'downloads' into 'files'
'Downloads' tab moved into new 'Files' tab which is also used to
contain the new uploads section consisting of similar file options.
